adding displayUserProduct function
