+++ src/pages/Profileadd.vue
@@ -26,27 +26,76 @@
     <p className="px-20 pb-0 text-left font-semibold text-5xl">User Profile</p>
   </div>
   <div className="flex flex-col items-start p-12 px-52">
+    <div className="flex">
+      <div className="flex flex-col items-start pb-8 pr-16">
+        <label
+          for="firstName"
+          class="text-gray-600 mt-3 h-auto w-52 text-left p-2"
+          >First Name</label
+        >
+        <input
+          class="border-4 border-gray-400 rounded-lg p-1.5"
+          type="text"
+          name="firstName"
+          placeholder="Enter your name"
+          v-model="firstName"
+        />
+      </div>
+      <div className="flex flex-col items-start pb-8">
+        <label
+          for="lastName"
+          class="text-gray-600 mt-3 h-auto w-52 text-left p-2"
+          >Last Name</label
+        >
+        <input
+          class="border-4 border-gray-400 rounded-lg p-1.5"
+          type="text"
+          name="lastName"
+          placeholder="Enter your name"
+          v-model="lastName"
+        />
+      </div>
+    </div>
+
     <div className="flex flex-col items-start pb-8">
-      <Textbox nametag="userName" text="Name" />
+      <!-- <Textbox nametag="aboutUser" text="About Me" /> -->
+      <label for="aboutMe" class="text-gray-600 mt-3 h-auto w-52 text-left p-2"
+        >About Me</label
+      >
+      <textarea
+        name="aboutMe"
+        class="border-4 border-gray-400 rounded-lg p-1.5"
+        id="aboutMe"
+        cols="60"
+        rows="5"
+        v-model="aboutMe"
+      ></textarea>
     </div>
-    <div className="flex flex-row gap-20 pb-8">
+    <div className="flex items-end">
       <div className="flex flex-col items-start">
-        <Textbox :name="name" text="Old Password" />
+        <!-- <Textbox nametag="shipAdd" text="Shipping Address" /> -->
+        <label
+          for="shippingAddress"
+          class="text-gray-600 mt-3 h-auto w-52 text-left p-2"
+          >Shipping Address</label
+        >
+        <textarea
+          name="shippingAddress"
+          class="border-4 border-gray-400 rounded-lg p-1.5"
+          id="shippingAddress"
+          cols="60"
+          rows="5"
+          v-model="shippingAddress"
+        ></textarea>
       </div>
-      <div className="flex flex-col items-start">
-        <Textbox :name="name" text="New Password" />
+      <div
+        class="p-2 border-2 border-black bg-blue-600 text-white rounded-lg ml-20 hover:bg-white hover:text-black transition ease-out duration-300"
+      >
+        <button @click="updateUser">Update Profile</button>
       </div>
-      <div className="flex flex-col items-start">
-        <Textbox :name="name" text="Retype New Password" />
-      </div>
-    </div>
-    <div className="flex flex-col items-start pb-8">
-      <Textbox nametag="aboutUser" text="About Me" />
-    </div>
-    <div className="flex flex-col items-start">
-      <Textbox nametag="shipAdd" text="Shipping Address" />
     </div>
   </div>
+
   <div>
     <div class="flex p-12 justify-around">
       <a className="pl-40 text-left text-4xl">My Products</a>
@@ -69,7 +118,7 @@
           @change="uploadProductImage($event)"
         />
         <label
-          class="border-2 border-gray-400 p-2 rounded-md"
+          class="border-2 border-gray-400 p-2 rounded-md hover:bg-blue-600 hover:text-white transition ease-linear duration-300"
           for="uploadProductImg"
           style="cursor: pointer"
           >Upload</label
@@ -82,37 +131,91 @@
           @click="popProductImage($event)"
         />
         <label
-          class="border-2 border-gray-400 p-2 rounded-md"
+          class="border-2 border-gray-400 p-2 rounded-md hover:bg-blue-600 hover:text-white transition ease-linear duration-300"
           for="popProductImage"
           style="cursor: pointer"
           >Delete</label
         >
       </div>
-      <!-- always create an array to store the src/data instead of using raw DOMs -->
     </div>
   </div>
-  <form @submit.prevent="handleSubmit">
-    <!-- need to add in require in the input -->
-    <!-- .prevent prevents the default submission of the form -->
-    <div class="flex flex-col items-start px-52">
-      <Textbox text="Name" name="Name" nametag="productName" />
-      <!-- <label for="name" className="mt-3 p-2 h-auto w-52 text-gr0 text-left">Name</label>
-      <input type="text" name="name" placeholder="Name" className="storybook-textbox" v-model="form.name" /> -->
-      <div class="flex">
-        <Textbox text="Points" name="Points" nametag="productPoints" />
-        <Dropdown class="pl-10 pt-7" />
-      </div>
-      <Textbox
-        text="Description"
-        name="Please enter item descriptions here."
-        nametag="productDescription"
-        size="large"
+  <div class="flex px-52">
+    <div class="flex flex-col items-start mr-16">
+      <label
+        for="productName"
+        class="text-gray-600 mt-3 h-auto w-52 text-left p-2"
+        >Product Name</label
+      >
+      <input
+        class="border-4 border-gray-400 rounded-lg p-1.5"
+        type="text"
+        name="productName"
+        placeholder="Enter product name"
+        v-model="productName"
+        required
       />
-      <div class="mt-7">
-        <Button type="submit" label="Add" :primary="true" />
-      </div>
     </div>
-  </form>
+    <div class="flex flex-col items-start mr-16">
+      <label
+        for="productPoints"
+        class="text-gray-600 mt-3 h-auto w-52 text-left p-2"
+        >Product Points</label
+      >
+      <input
+        class="border-4 border-gray-400 rounded-lg p-1.5"
+        type="number"
+        name="productPoints"
+        placeholder="Enter product value"
+        v-model="productPoints"
+        required
+      />
+    </div>
+    <div class="flex flex-col items-start mr-16">
+      <label
+        for="productQty"
+        class="text-gray-600 mt-3 h-auto w-52 text-left p-2"
+        >Product Quantity</label
+      >
+      <input
+        class="border-4 border-gray-400 rounded-lg p-1.5"
+        type="number"
+        name="productQty"
+        placeholder="Enter product quantity"
+        v-model="productQty"
+        required
+      />
+    </div>
+    <div class="flex justify-start">
+      <Dropdown class="" />
+    </div>
+  </div>
+  <div class="flex flex-col items-start px-52">
+    <label
+      for="productDescription"
+      class="text-gray-600 mt-3 h-auto w-52 text-left p-2"
+      >Description</label
+    >
+    <textarea
+      name="productDescription"
+      class="border-4 border-gray-400 rounded-lg p-1.5"
+      id="productDescription"
+      cols="90"
+      rows="10"
+      placeholder="Enter product description here."
+      v-model="productDescription"
+    ></textarea>
+    <div class="mt-7">
+      <!-- <Button class="mb-10 hover:bg-white hover:text-black hover:border-black transition ease-linear duration-500" type="submit" label="Add Product" :primary="true" /> -->
+      <!-- <button type="submit" class="border-2 p-3 rounded-lg bg-blue-400 border-gray-300 text-white mb-28">Add Product</button> -->
+      <Button
+        class="mb-10 transform hover:scale-125 hover:opacity-75 transition ease-out duration-300"
+        type="submit"
+        @click="handleSubmit"
+        label="Add Product"
+        :primary="true"
+      />
+    </div>
+  </div>
 </template>
 
 <script>
@@ -122,15 +225,21 @@ import Image from "/@/components/molecule/Image/Image.vue";
 import Dropdown from "/@/components/molecule/Dropdown/Dropdown.vue";
 import Button from "/@/components/molecule/Button/Button.vue";
 import { getCurrentInstance, onBeforeMount, reactive, ref } from "vue";
-import { createProduct, currentUser } from "/@/utils/firebase";
-import router from '../router/router';
-
+import {
+  createProduct,
+  currentUser,
+  getUserProfile,
+  getUserProfileDoc,
+  updateUserProfile,
+} from "/@/utils/firebase";
+import firebase from "firebase";
 
 export default {
-  name: "Profileadd",
+  name: "ProfileAdd",
   data() {
     return {
-      name: "*********",
+      firstName: "",
+      lastName: "",
       images: [
         {
           id: "1",
@@ -144,42 +253,79 @@ export default {
       productName: "",
       productPoints: "",
       productDescription: "",
+      productQty: "",
       condition: "",
-      // userName: "",
-      // shipAdd: "",
-      // aboutUser: "",
+      shippingAddress: "",
+      aboutMe: "",
     };
   },
+  created() {
+    firebase.auth().onAuthStateChanged(async (user) => {
+      if (user) {
+        console.log("user is logged in: ", user);
+        console.log(user.displayName);
+        const userDetails = getUserProfileDoc(user.email);
+        if ((await userDetails).empty) {
+          console.log("no matching user");
+        } else {
+          (await userDetails).forEach((doc) => {
+            const details = doc.data();
+            this.aboutMe = details.about;
+            this.shippingAddress = details.address;
+            this.firstName = details.first_name;
+            this.lastName = details.last_name;
+          });
+        }
+      } else {
+        console.log("user is logged out.");
+      }
+    });
+    getUserProfile();
+  },
   methods: {
+    async updateUser() {
+      const user = currentUser();
+      if (user) {
+        await updateUserProfile(user.uid, this.aboutMe, this.shippingAddress, this.firstName, this.lastName);
+        console.log("successfully updated user profile!");
+      } else {
+        console.log("failed to update user profile...");
+      }
+      
+    },
     handleBack() {
       this.$router.go(-1);
     },
-    handleSubmit(event) {
+    handleSubmit() {
       // takes the data from the form and put it into a suitable format using
       // Object.fromentries to grab individual input and destructure it
       // name attribute must be given in each input
       let testProduct;
-      const {
-        productName,
-        productPoints,
-        condition,
-        productDescription,
-      } = Object.fromEntries(new FormData(event.target));
-      this.productName = productName;
-      this.productPoints = productPoints;
-      this.condition = condition;
-      this.productDescription = productDescription;
+      const user = currentUser();
+      // const {
+      //   productName,
+      //   productPoints,
+      //   condition,
+      //   productDescription,
+      // } = Object.fromEntries(new FormData(event.target));
+      // this.productName = productName;
+      // this.productPoints = productPoints;
+      // this.condition = condition;
+      // this.productDescription = productDescription;
       // to see the values generated
       console.log(
         this.productName,
         this.productPoints,
+        this.productQty,
         this.condition,
         this.productDescription
       );
       this.addProductDetails.push({
         id: Date.now(),
+        uploadedBy: user.uid,
         name: this.productName,
         points: this.productPoints,
+        quanity: this.productQty,
         conditions: this.condition,
         description: this.productDescription,
         status: "true",
@@ -190,8 +336,10 @@ export default {
       // adding objects into testProduct variable
       testProduct = {
         id: Date.now(),
+        uploadedBy: user.uid,
         name: this.productName,
         points: this.productPoints,
+        quanity: this.productQty,
         conditions: this.condition,
         description: this.productDescription,
         status: "true",
@@ -201,27 +349,26 @@ export default {
       // for status, true means product is available
       // false means product is in shipping, shipped or paid
       // const user = currentUser();
-      const user = currentUser();
+
       if (user !== null) {
-        const uid = user.uid;
         // const newProduct = db.collection("userProfile").doc(uid).collection('products').add(this.addProductDetails)
         // create new document with info from testProduct into the subCollection
-        createProduct(testProduct, uid)
+        createProduct(testProduct, user.uid)
           .then(function () {
             console.log("product created.");
+            // clear the inputs after adding product
           })
           .then(function () {
             alert("Successfully added the product");
-          })
+          });
       }
-      // clear the inputs after adding product
-      // this.productName = "";
-      // this.productPoints = "";
-      // this.condition = "almostNew";
-      // this.productDescription = "";
-      this.$router.push('/profile_add')
+      this.productName = "";
+      this.productPoints = "";
+      this.condition = "";
+      this.productDescription = "";
+      this.productQty = "";
       this.images = [];
-
+      // this.$router.push("/profile_add");
       // need create a storage link to store images
       // link the user ID with their products
     },
@@ -248,13 +395,6 @@ export default {
     },
   },
   setup() {
-    // const user = firebase.auth().currentUser;
-    // if (user) {
-    //   this.shipAdd = user.shippingAddress;
-    //   this.userName = user.name;
-    //   this.aboutUser = user.aboutMe;
-    // }
-
     onBeforeMount(() => {
       const user = currentUser();
       // user display null...maybe something is missing
@@ -277,7 +417,7 @@ export default {
       // shipAdd,
       // usersName,
       // aboutUser
-    }
+    };
     //   // handleSubmit: async () => {
     //   //   await createProfile({...form})
     //   //   form.name = ''
